addressed remaining review comments
