+++ gqcp/tests/Basis/TransformationMatrix_test.cpp
@@ -61,12 +61,12 @@ BOOST_AUTO_TEST_CASE ( gamessUS ) {
           1.0, 3.0;
 
     GQCP::TransformationMatrix<double> T2 (6);
-	T2 <<-1.0, 0.0,-4.0, 5.0,-7.0, 8.0,
-		  1.0, 3.0,-4.0, 5.0,-7.0, 8.0,
-		  0.0, 0.0,-4.0, 5.0,-7.0, 8.0,
-		  1.0, 9.0,-4.0, 5.0,-7.0, 8.0,
-		  0.0, 0.0,-4.0, 5.0,-7.0, 8.0,
-		  1.0,-7.0,-4.0, 5.0,-7.0, 8.0;
+	T2 << -1.0, 0.0,-4.0, 5.0,-7.0, 8.0,
+		   1.0, 3.0,-4.0, 5.0,-7.0, 8.0,
+		   0.0, 0.0,-4.0, 5.0,-7.0, 8.0,
+		   1.0, 9.0,-4.0, 5.0,-7.0, 8.0,
+		   0.0, 0.0,-4.0, 5.0,-7.0, 8.0,
+		   1.0,-7.0,-4.0, 5.0,-7.0, 8.0;
 		
 	// Test them against a reference representation
     BOOST_CHECK(reference_T1.compare(T1.asGAMESSUSVECGroup()) == 0);
